requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
gspread>=6.0.0
google-auth>=2.27.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
